matcher/pattern/file/serialized: move errors into their own module
